djangoApi/app/views.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status
from .models import Puppy
from .serializers import PuppySerializer


import json
import requests
from django.shortcuts import render, HttpResponse

# Create your views here.
def index(request):
    return HttpResponse('Hello World!')

def indexer(request):
    return HttpResponse('Hello eyin temiii!')

def number_two(request):
    return HttpResponse('My second view lives here')

@api_view(['GET', 'DELETE', 'PUT'])
def get_delete_update_puppy(request, pk):
    try:
        puppy = Puppy.objects.get(pk=pk)
    except Puppy.DoesNotExist:
        return Response(status=status.HTTP_404_NOT_FOUND)

    # get details of a single puppy
    if request.method == 'GET':
        serializer = PuppySerializer(puppy)
        return Response(serializer.data)

    # update details of a single puppy
    if request.method == 'PUT':
        serializer = PuppySerializer(puppy, data=request.data)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)

    # delete a single puppy
    if request.method == 'DELETE':
        puppy.delete()
        return Response(status=status.HTTP_204_NO_CONTENT)


@api_view(['GET', 'POST'])
def get_post_puppies(request):
    # get all puppies
    if request.method == 'GET':
        puppies = Puppy.objects.all()
        serializer = PuppySerializer(puppies, many=True)
        return Response(serializer.data)
    # insert a new record for a puppy
    if request.method == 'POST':
        data = {
            'name': request.data.get('name'),
            'age': int(request.data.get('age')),
            'breed': request.data.get('breed'),
            'color': request.data.get('color')
        }
        serializer = PuppySerializer(data=data)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data, status=status.HTTP_201_CREATED)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)

def getUser(request):
    json_list = []
    req = requests.get('https://abokifx.com/')
    print("Here comes", req)
    json_list.append(json.loads(req.content))
    parsed_data = []
    user_data = {}
    for data in json_list:
      user_data['name'] = data['name']
      user_data['blog'] = data['blog']
      user_data['email'] = data['email']
      user_data['public_gists'] = data['public_gists']
      user_data['public_repos'] = data['public_repos']
      user_data['avaar_url'] = data['avatar_url']
      user_data['followers'] = data['followers']
      user_data['following'] = data['following']
    parsed_data.append(user_data)
    return render(request, 'app/profile.html', {'data': parsed_data})

