djangoApi/requirements.txt
autopep8==1.3.5
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
djangorestframework==3.8.2
idna==2.7
psycopg2==2.7.5
pycodestyle==2.4.0
pytz==2018.4
requests==2.19.1
urllib3==1.23
